Rename flags.ts to options.ts
