+++ packages/openapi-generator/src/options/index.ts
@@ -1,4 +1,4 @@
 export * from './generator-options';
 export * from './options-per-service';
 export * from './tsconfig-json';
-export * from './flags';
+export * from './options';
