+++ packages/openapi-generator/src/options/options.ts
